refactoring directives module: breaking up big files
